+++ pysat/instruments/pysat_ndtesting.py
@@ -15,8 +15,10 @@ name = 'ndtesting'
 
 pandas_format = False
 tags = {'': 'Regular testing data set'}
-inst_ids = {'': ['']}
-_test_dates = {'': {'': dt.datetime(2009, 1, 1)}}
+inst_ids = {'': [tag for tag in tags.keys()]}
+_test_dates = {'': {tag: dt.datetime(2009, 1, 1) for tag in tags.keys()}}
+_test_load_opt = {'': {'': [{'num_extra_time_coords': 0},
+                            {'num_extra_time_coords': 1}]}}
 
 epoch_name = u'time'
 
@@ -26,13 +28,15 @@ init = mm_test.init
 # Clean method
 clean = mm_test.clean
 
-# Optional method, preprocess
+# Optional methods
+concat_data = mm_test.concat_data
 preprocess = mm_test.preprocess
 
 
 def load(fnames, tag='', inst_id='', non_monotonic_index=False,
          non_unique_index=False, malformed_index=False, start_time=None,
-         num_samples=864, test_load_kwarg=None, max_latitude=90.):
+         num_samples=864, test_load_kwarg=None, max_latitude=90.0,
+         num_extra_time_coords=0):
     """Load the test files.
 
     Parameters
@@ -65,7 +69,9 @@ def load(fnames, tag='', inst_id='', non_monotonic_index=False,
         correctly. (default=None)
     max_latitude : float
         Latitude simulated as `max_latitude` * cos(theta(t))`, where
-        theta is a linear periodic signal bounded by [0, 2 * pi) (default=90.).
+        theta is a linear periodic signal bounded by [0, 2 * pi) (default=90.0)
+    num_extra_time_coords : int
+        Number of extra time coordinates to include. (default=0)
 
     Returns
     -------
@@ -165,37 +171,60 @@ def load(fnames, tag='', inst_id='', non_monotonic_index=False,
                                    dtype=np.int64))
 
     # Add dummy coords
-    data.coords['x'] = (('x'), np.arange(17))
-    data.coords['y'] = (('y'), np.arange(17))
-    data.coords['z'] = (('z'), np.arange(15))
+    data.coords['x'] = (('x'), np.arange(7))
+    data.coords['y'] = (('y'), np.arange(7))
+    data.coords['z'] = (('z'), np.arange(5))
+
+    # Add extra time coords
+    for i in range(num_extra_time_coords):
+        ckey = 'time{:d}'.format(i)
+        tindex = data.indexes[epoch_name][:-1 * (i + 1)]
+        data.coords[ckey] = (
+            (ckey), [itime + dt.timedelta(microseconds=1 + i)
+                     for i, itime in enumerate(tindex)])
 
     # Create altitude 'profile' at each location to simulate remote data
     num = len(data['uts'])
     data['profiles'] = (
         (epoch_name, 'profile_height'),
-        data['dummy3'].values[:, np.newaxis] * np.ones((num, 15)))
-    data.coords['profile_height'] = ('profile_height', np.arange(15))
+        data['dummy3'].values[:, np.newaxis] * np.ones(
+            (num, data.coords['z'].shape[0])))
+    data.coords['profile_height'] = ('profile_height',
+                                     np.arange(len(data.coords['z'])))
 
     # Profiles that could have different altitude values
     data['variable_profiles'] = (
         (epoch_name, 'z'), data['dummy3'].values[:, np.newaxis]
-        * np.ones((num, 15)))
+        * np.ones((num, data.coords['z'].shape[0])))
     data.coords['variable_profile_height'] = (
-        (epoch_name, 'z'), np.arange(15)[np.newaxis, :] * np.ones((num, 15)))
+        (epoch_name, 'z'), np.arange(data.coords['z'].shape[0])[np.newaxis, :]
+        * np.ones((num, data.coords['z'].shape[0])))
 
     # Create fake image type data, projected to lat / lon at some location
     # from satellite.
     data['images'] = ((epoch_name, 'x', 'y'),
                       data['dummy3'].values[
-                          :, np.newaxis, np.newaxis] * np.ones((num, 17, 17)))
-    data.coords['image_lat'] = \
-        ((epoch_name, 'x', 'y'),
-         np.arange(17)[np.newaxis,
-                       np.newaxis,
-                       :] * np.ones((num, 17, 17)))
+                          :, np.newaxis, np.newaxis]
+                      * np.ones((num, data.coords['x'].shape[0],
+                                 data.coords['y'].shape[0])))
+    data.coords['image_lat'] = ((epoch_name, 'x', 'y'),
+                                np.arange(data.coords['x'].shape[0])[
+                                    np.newaxis, np.newaxis, :]
+                                * np.ones((num, data.coords['x'].shape[0],
+                                           data.coords['y'].shape[0])))
     data.coords['image_lon'] = ((epoch_name, 'x', 'y'),
-                                np.arange(17)[np.newaxis, np.newaxis,
-                                              :] * np.ones((num, 17, 17)))
+                                np.arange(data.coords['x'].shape[0])[
+                                    np.newaxis, np.newaxis, :]
+                                * np.ones((num, data.coords['x'].shape[0],
+                                           data.coords['y'].shape[0])))
+
+    # There may be data that depends on alternate time indices
+    for i in range(num_extra_time_coords):
+        alt_epoch = 'time{:d}'.format(i)
+        data['variable_profiles{:d}'.format(i)] = (
+            (alt_epoch, 'z'), np.full(shape=(data.coords[alt_epoch].shape[0],
+                                             data.coords['z'].shape[0]),
+                                      fill_value=100.0 + i))
 
     meta = mm_test.initialize_test_meta(epoch_name, data.keys())
     return data, meta